fix(fe): chat header is sticky and transparent
